Filter for test env to only execute relevant test based on env.
