+++ .gitignore
@@ -5,4 +5,3 @@
 .vscode/ipch
 .vscode/extensions.json
 .vscode/settings.json
-.vscode/extensions.json
