+++ platformio.ini
@@ -12,8 +12,10 @@
 platform = atmelavr
 board = megaatmega2560
 framework = arduino
+test_filter = target/*
 ;lib_deps = throwtheswitch/Unity@^2.5.2
 
 [env:native]
 platform = native
+test_filter = native/*
 build_flags = -D WINDOWS_TEST
